Scale over 128 to prevent rounding issues
Bench: 7373334

+++ src/tune.h
@@ -80,7 +80,7 @@ typedef struct {
   int8_t phase;
   int8_t stm;
   float result;
-  float scale;
+  int scale;
   float phaseMg;
   float phaseEg;
   Score staticEval;
